Relocate cloneObjectEntries into project-index utilities
